feat(featureactions): move feature-specific code to featureactionmanager
cleaned up typedef; moved feature-specific code from importactionmanager to featureactionmanager
