+++ src/os/im/action/importactioncallbackconfig.js
@@ -1,0 +1,15 @@
+goog.module('os.im.action.ImportActionCallbackConfig');
+
+/**
+ * @typedef {{
+ *  color: (Array<Iterable>|undefined),
+ *  labelUpdateShown: boolean,
+ *  notifyStyleChange: boolean,
+ *  setColor: boolean,
+ *  setFeaturesStyle: boolean
+ * }}
+ */
+let ImportActionCallbackConfig;
+
+
+exports = ImportActionCallbackConfig;
